feat : deleteStudyroom 메서드 구현 - 예외처리와 로그 강화
